refactor join functions
