+++ bodo/libs/array.py
@@ -2784,182 +2784,6 @@ def get_null_shuffle_info(typingctx):
 
 
 @intrinsic
-def hash_join_table(
-    typingctx,
-    left_table_t,
-    right_table_t,
-    left_parallel_t,
-    right_parallel_t,
-    n_keys_t,
-    n_data_left_t,
-    n_data_right_t,
-    same_vect_t,
-    key_in_out_t,
-    same_need_typechange_t,
-    is_left_t,
-    is_right_t,
-    is_join_t,
-    extra_data_col_t,
-    indicator,
-    _bodo_na_equal,
-    _bodo_rebalance_output_if_skewed,
-    cond_func,
-    left_col_nums,
-    left_col_nums_len,
-    right_col_nums,
-    right_col_nums_len,
-    num_rows_ptr_t,
-):
-    """
-    Interface to the hash join of two tables.
-    """
-    assert left_table_t == table_type
-    assert right_table_t == table_type
-
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),
-            [
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(64),
-                lir.IntType(64),
-                lir.IntType(64),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(64),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(64),
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="hash_join_table"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        table_type(
-            left_table_t,
-            right_table_t,
-            types.boolean,
-            types.boolean,
-            types.int64,
-            types.int64,
-            types.int64,
-            types.voidptr,
-            types.voidptr,
-            types.voidptr,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.voidptr,
-            types.voidptr,
-            types.int64,
-            types.voidptr,
-            types.int64,
-            types.voidptr,
-        ),
-        codegen,
-    )
-
-
-@intrinsic
-def nested_loop_join_table(
-    typingctx,
-    left_table_t,
-    right_table_t,
-    left_parallel_t,
-    right_parallel_t,
-    is_left_t,
-    is_right_t,
-    key_in_output_t,
-    need_typechange_t,
-    _bodo_rebalance_output_if_skewed,
-    cond_func,
-    left_col_nums,
-    left_col_nums_len,
-    right_col_nums,
-    right_col_nums_len,
-    num_rows_ptr_t,
-):
-    """
-    Call cpp function for cross join of two tables.
-    """
-    assert left_table_t == table_type, "nested_loop_join_table: cpp table type expected"
-    assert right_table_t == table_type, (
-        "nested_loop_join_table: cpp table type expected"
-    )
-
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),
-            [
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(1),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(1),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(64),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(64),
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="nested_loop_join_table"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        table_type(
-            left_table_t,
-            right_table_t,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.boolean,
-            types.voidptr,
-            types.voidptr,
-            types.boolean,
-            types.voidptr,
-            types.voidptr,
-            types.int64,
-            types.voidptr,
-            types.int64,
-            types.voidptr,
-        ),
-        codegen,
-    )
-
-
-@intrinsic
 def shuffle_renormalization(typingctx, table_t, random_t, random_seed_t, is_parallel_t):
     """
     Interface to the rebalancing of the table
